requirements.txt

torch==1.9.0
pytorch-lightning==1.4.9
fastapi==0.68.0
uvicorn==0.15.0
